refactor(bitbucket): extract evidence mappers into _evidence.py
Move map_list_bitbucket_commits, map_get_bitbucket_file_contents, and
map_search_bitbucket_code out of their tools' __init__.py facades into
dedicated sibling modules, per review feedback that facade files
should only expose the public tool interface.
